Add docker and docker-compose files
Allow the app to run over a docker container the database is named inside of docker-compose.yml.

In order to don't need to install everything to be up and runing for tests, the user just need to run sudo docker-compose build && sudo docker-compose up. This is usefull to who just want to run the application in local machine.